+++ processSlice_pyrex_py2.pyx
@@ -2,10 +2,10 @@
 
 #  A Pyrex extensions to calculate a slice of the overall calculation of Pi using quadrature.
 #
-#  Copyright © 2010 Russel Winder
+#  Copyright © 2010--2011 Russel Winder
 
-def processSlice ( long id , long sliceSize , double delta ) :
-  cdef long i
+def processSlice ( int id , int sliceSize , double delta ) :
+  cdef int i
   cdef double sum , x
   sum = 0.0
   for i in xrange ( 1 + id * sliceSize , ( id + 1 ) * sliceSize + 1 ) :
